+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMediaAnimation.kt
@@ -22,11 +22,8 @@ data class InputMediaAnimation(
     override val type: String = "animation"
 
     @SerialName(mediaField)
-    val media: String
-        get() = file.let {
-            when (it) {
-                is FileId -> it.fileId
-                is MultipartFile -> it.fileId.toInputMediaFileAttachmentName()
-            }
-        }
+    val media: String = when (file) {
+        is FileId -> file.fileId
+        is MultipartFile -> file.fileId.toInputMediaFileAttachmentName()
+    }
 }