+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/InputMediaAudio.kt
@@ -23,11 +23,8 @@ data class InputMediaAudio(
     override val type: String = "audio"
 
     @SerialName(mediaField)
-    val media: String
-        get() = file.let {
-            when (it) {
-                is FileId -> it.fileId
-                is MultipartFile -> it.fileId.toInputMediaFileAttachmentName()
-            }
-        }
+    val media: String = when (file) {
+        is FileId -> file.fileId
+        is MultipartFile -> file.fileId.toInputMediaFileAttachmentName()
+    }
 }